% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cbs_data}
\alias{cbs_data}
\alias{gemeenten}
\alias{ggdregios}
\alias{jeugdregios}
\alias{nuts3regios}
\alias{postcodes4}
\alias{provincies}
\alias{veiligheidsregios}
\title{Data Sets with Geometries of Dutch Provinces, Municipalities and Zipcodes}
\format{
A \code{sf}/\link{data.frame} with 3 variables:
\itemize{
\item \code{...}\cr column name of the identifier, which is the singular form of the name of the data set (i.e., \code{gemeenten$gemeente})
\item \code{area_km2}\cr area in square kilometres
\item \code{geometry}\cr the polygonal form of the area
}

An object of class \code{sf} (inherits from \code{data.frame}) with 25 rows and 3 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 42 rows and 3 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 40 rows and 3 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 1060 rows and 3 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 12 rows and 3 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 25 rows and 3 columns.
}
\usage{
gemeenten

ggdregios

jeugdregios

nuts3regios

postcodes4

provincies

veiligheidsregios
}
\description{
Data Sets with Geometries of Dutch Provinces, Municipalities and Zipcodes
}
\details{
All data sets (maps) contain all of the Netherlands, except for \link{postcodes4} which only contains the northern Netherlands.

See \href{https://github.com/certe-medical-epidemiology/certedata/blob/main/data-raw/update_gis.R}{the repository file} to update these data sets.
}
\keyword{datasets}
