% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adherence_area.R
\name{adherence_area}
\alias{adherence_area}
\title{Determine Hospital Adherence Area}
\usage{
adherence_area(zipcode)
}
\arguments{
\item{zipcode}{A character or numeric vector of zipcodes. Six-character
zipcodes are automatically truncated to the first 4 digits using
\code{substr()}.}
}
\value{
A character vector of hospital names, the same length as \code{zipcode}.
}
\description{
Looks up the nearest hospital for a given zipcode (4 digits) based on
the shortest distance in the \link{postcodes4_afstanden} dataset.
}
\details{
The function looks up the distance from each input zipcode to all
hospital zipcodes in \link{postcodes4_afstanden}. The hospital with the
shortest distance is returned. In case of equal distances, the first
hospital in definition order is chosen.

The following hospitals and their zipcodes are hard-coded:\tabular{ll}{
   Hospital \tab Zip Code \cr
   Antonius \tab 8601 \cr
   Frisius Heerenveen \tab 8441 \cr
   Frisius Leeuwarden \tab 8934 \cr
   Martini \tab 9728 \cr
   Nij Smellinghe \tab 9202 \cr
   Ommelander \tab 9679 \cr
   Treant \tab 7824 \cr
   Wilhelmina \tab 9401 \cr
}


Zipcodes not found in \link{postcodes4_afstanden} return \code{NA}.
}
\examples{
adherence_area("9700")
adherence_area(c("9713", "7702", "8896"))

# six-character zipcodes are automatically truncated
adherence_area("9251AB")

if (requireNamespace("plot2", quietly = TRUE)) {
  library(certeplot2)
  geo_postcodes4 |>
     crop_certe() |>
     mutate(adherence = suppressMessages(adherence_area(postcode)),
            .after = 1) |>
     plot2(category = adherence) |>
     add_sf(geo_provincies |> crop_certe(),
            colour_fill = NA, colour = "black", linewidth = 0.5)
}
}
