% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adherence_area.R
\name{adherence_area}
\alias{adherence_area}
\title{Determine (Hospital) Adherence Area}
\usage{
adherence_area(
  zipcode,
  reference = c(Antonius = "8601", `Frisius Heerenveen` = "8441", `Frisius Leeuwarden` =
    "8934", Martini = "9728", `Nij Smellinghe` = "9202", Ommelander = "9679", Treant =
    "7824", Wilhelmina = "9401")
)
}
\arguments{
\item{zipcode}{A character or numeric vector of zipcodes. Six-character zipcodes are automatically truncated to the first 4 digits using \code{substr()}.}

\item{reference}{A character vector of zipcodes to use as reference, preferably named. Defaults to hospitals in the region.}
}
\value{
A character vector of names of \code{reference}, the same length as \code{zipcode}.
}
\description{
Looks up the nearest reference name for a given zipcode (4 digits) based on the shortest distance in the \link{postcodes4_afstanden} dataset.
}
\details{
The function looks up the distance from each input zipcode to all zipcodes of \code{reference} in \link{postcodes4_afstanden}. The names with the shortest distance is returned. In case of equal distances, the first name in definition order is chosen.

Zipcodes not found in \link{postcodes4_afstanden} return \code{NA}.
}
\examples{
adherence_area("9700")
adherence_area(c("9713", "7702", "8896"))

# six-character zipcodes are automatically truncated
adherence_area("9251AB")

if (requireNamespace("certeplot2", quietly = TRUE)) {
  geo_postcodes4 |>
     crop_certe() |>
     plot2(category = suppressMessages(adherence_area(postcode)),
           datalabels = FALSE) |>
     plot2::add_sf(geo_provincies |> crop_certe(),
            colour_fill = NA, colour = "black", linewidth = 0.5)
}

if (requireNamespace("certeplot2", quietly = TRUE)) {
  hospitals <- c("Antonius"           = "8601",
                 "Frisius Heerenveen" = "8441",
                 "Frisius Leeuwarden" = "8934",
                 "Martini"            = "9728",
                 "Nij Smellinghe"     = "9202",
                 "Ommelander"         = "9679",
                 "Treant"             = "7824",
                 "Wilhelmina"         = "9401")
  zip4 <- geo_postcodes4 |> crop_certe()

  zip4 |>
    plot2(category = adherence_area(postcode, reference = hospitals),
          datalabels = FALSE) |>
    plot2::add_sf(geo_provincies |> crop_certe(),
           colour_fill = "#FFFFFF77", colour = "black", linewidth = 0.5) |>
    plot2::add_sf(zip_to_sf(hospitals),
           datalabels = names(hospitals),
           colour = "black")
}
}
