% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cbs_data}
\alias{cbs_data}
\alias{geo_postcodes4}
\alias{geo_postcodes3}
\alias{geo_postcodes2}
\alias{geo_ggdregios}
\alias{geo_nuts3}
\alias{geo_gemeenten}
\alias{geo_provincies}
\title{Data Sets with Geometries of Dutch Provinces, Municipalities and Zipcodes}
\format{
A \code{sf}/\link{data.frame} with 3 variables:
\itemize{
\item \code{...}\cr column name of the identifier, which is the singular form of the name of the data set (i.e., \code{geo_gemeenten$gemeente})
\item \code{inwoners}\cr number of inhabitants
\item \code{oppervlakte_km2}\cr area in square kilometres
\item \code{geometry}\cr the multipolygonal form of the area
}

An object of class \code{sf} (inherits from \code{data.frame}) with 799 rows and 4 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 97 rows and 4 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 25 rows and 4 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 40 rows and 4 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 345 rows and 4 columns.

An object of class \code{sf} (inherits from \code{data.frame}) with 12 rows and 4 columns.
}
\usage{
geo_postcodes4

geo_postcodes3

geo_postcodes2

geo_ggdregios

geo_nuts3

geo_gemeenten

geo_provincies
}
\description{
Data Sets with Geometries of Dutch Provinces, Municipalities and Zipcodes
}
\details{
See \href{https://github.com/certe-medical-epidemiology/certegis/blob/main/data-raw/update_gis.R}{the repository file} to update these data sets.
}
\keyword{datasets}
