% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoding.R
\name{geocoding}
\alias{geocoding}
\alias{geocode}
\alias{reverse_geocode}
\title{Geocoding to Find Coordinates and Addresses}
\source{
Data © OpenStreetMap contributors, ODbL 1.0. \url{https://osm.org/copyright}
}
\usage{
geocode(place, as_coordinates = FALSE, only_netherlands = TRUE)

reverse_geocode(sf_data)
}
\arguments{
\item{place}{a (vector of) names or addresses of places}

\item{as_coordinates}{a \link{logical} to indicate whether the result should be returned as coordinates (i.e., class \code{sfc_POINT})}

\item{only_netherlands}{a \link{logical} to indicate whether only Dutch places should be searched}

\item{sf_data}{an 'sf' object or an 'sfc' object (i.e., a vector with geometric \code{sfc_POINT}s). Can also be a character vector, in which case \code{\link[=get_coordinates]{get_coordinates()}} will be called first.}
}
\description{
\strong{Geocoding} is the process of retrieving geographic coordinates based on text, such as an address or the name of a place (\href{https://en.wikipedia.org/wiki/Address_geocoding}{Wikipedia page}). On the other hand, \strong{reverse geocoding} is the process of retrieving the name and address from geographic coordinates (\href{https://en.wikipedia.org/wiki/Reverse_geocoding}{Wikipedia page}).
}
\details{
These functions use \href{https://openstreetmap.org}{OpenStreetMap (OSM)}.

\code{\link[=geocode]{geocode()}} provides geocoding and returns an 'sf' \link{data.frame} at default. In case of multiple results, the distance from the main Certe building in Groningen is leading.

\code{\link[=reverse_geocode]{reverse_geocode()}} provides reversed geocoding and returns a \link{data.frame} with the columns "name", "address", "zipcode" and "city".

For both functions, the OSM API will only be called on unique input values, to increase speed.
}
\examples{
# geocoding: retrieve 'sf' data.frame based on place names
coord <- geocode("Van Swietenlaan 2, Groningen")
coord

# reverse geocoding: get the name and address
reverse_geocode(coord)

# places can be any text, and the results are prioritised based on
# the distance from the main Certe building, so:
reverse_geocode(c("Certe", "IKEA"))

hospitals <- geocode(c("Martini ziekenhuis",
                       "Medisch Centrum Leeuwarden",
                       "Tjongerschans Heerenveen",
                       "Scheper Emmen"))
hospitals

if (require("certeplot2")) {
  geo_gemeenten \%>\%
    crop_certe() \%>\%
    plot2(datalabels = FALSE) \%>\%
    add_sf(hospitals, colour = "certeroze", datalabels = place)
}
}
