% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{add_map}
\alias{add_map}
\alias{is.sf}
\alias{cut_northern_nl}
\alias{filter_sf}
\alias{filter_geolocation}
\alias{latitude}
\alias{longitude}
\title{Geodata Functions}
\usage{
add_map(data, maptype = "postcodes4", by = NULL, cut_northern_nl = TRUE, ...)

is.sf(sf_data)

cut_northern_nl(sf_data)

filter_sf(sf_data, xmin = NULL, xmax = NULL, ymin = NULL, ymax = NULL)

filter_geolocation(sf_data, ..., col_zipcode = NULL)

latitude(sf_data)

longitude(sf_data)
}
\arguments{
\item{data}{data to join left to the geodata}

\item{maptype}{type of geometric data, must be one of: \code{"gemeenten"}, \code{"ggdregios"}, \code{"inwoners_per_postcode_leeftijd"}, \code{"jeugdregios"}, \code{"nuts3regios"}, \code{"postcodes"}, \code{"postcodes4"}, \code{"postcodes4_afstanden"}, \code{"provincies"}, \code{"veiligheidsregios"}}

\item{sf_data}{a data set of class 'sf'}

\item{xmin, xmax, ymin, ymax}{coordination filters for \code{sf_data}}
}
\value{
An \code{sf} model. The column with geodata is always called \code{"geometry"}.
}
\description{
Geodata Functions
}
\details{
\code{\link[=cut_northern_nl]{cut_northern_nl()}} cuts any geometry on the northern three provinces of the Netherlands.

\code{\link[=filter_sf]{filter_sf()}} filters an sf object on coordinates, and is internally used by \code{\link[=cut_northern_nl]{cut_northern_nl()}}.

\code{\link[=latitude]{latitude()}} and \code{\link[=longitude]{longitude()}} return these specific geographic properties of \code{sf_data}.
}
\examples{
if (require("certeplot2")) {

  postcodes4 \%>\% 
    filter_geolocation(gemeente == "Tytsjerksteradiel") \%>\% 
    plot2()

}
}
