% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{get_map}
\alias{get_map}
\alias{add_map}
\alias{is.sf}
\alias{crop_certe}
\alias{filter_sf}
\alias{filter_geolocation}
\alias{latitude}
\alias{longitude}
\title{Geodata Functions}
\usage{
get_map(maptype = "postcodes4")

add_map(data, maptype = NULL, by = NULL, crop_certe = TRUE)

is.sf(sf_data)

crop_certe(sf_data)

filter_sf(sf_data, xmin = NULL, xmax = NULL, ymin = NULL, ymax = NULL)

filter_geolocation(sf_data, ..., col_zipcode = NULL)

latitude(sf_data)

longitude(sf_data)
}
\arguments{
\item{maptype}{type of geometric data, must be one of: \code{"gemeenten"}, \code{"ggdregios"}, \code{"nuts3"}, \code{"postcodes2"}, \code{"postcodes3"}, \code{"postcodes4"}, \code{"provincies"}. For \code{\link[=add_map]{add_map()}}, this is determined automatically if left blank.}

\item{data}{data set to join left to the geodata}

\item{by}{column to join by}

\item{crop_certe}{\link{logical} to keep only the Certe region}

\item{sf_data}{a data set of class 'sf'}

\item{xmin, xmax, ymin, ymax}{coordination filters for \code{sf_data}}

\item{...}{filters to set}

\item{col_zipcode}{column with zip codes}
}
\value{
An \code{sf} model. The column with geodata is always called \code{"geometry"}.
}
\description{
These are function to work with geographical data.
}
\details{
\code{\link[=crop_certe]{crop_certe()}} cuts any geometry on the Certe region (more or less the Northern three provinces of the Netherlands).

\code{\link[=filter_sf]{filter_sf()}} filters an sf object on coordinates, and is internally used by \code{\link[=crop_certe]{crop_certe()}}.

\code{\link[=latitude]{latitude()}} and \code{\link[=longitude]{longitude()}} return these specific geographic properties of \code{sf_data}.
}
\examples{
library(sf)

get_map() # defaults to the geo_postcodes4 data set

data.frame(postcode = 7702, number_of_cases = 3) \%>\% 
  add_map()

geo_provincies \%>\% crop_certe()

# filter on a latitude of 52.5 degrees and higher
geo_provincies \%>\% filter_sf(ymin = 52.5)

if (require("certeplot2")) {

  geo_postcodes4 \%>\% 
    filter_geolocation(gemeente == "Tytsjerksteradiel") \%>\% 
    plot2(category = inwoners,
          datalabels = postcode)

}

latitude(geo_provincies)
longitude(geo_provincies)
}
