% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{postcodes4_afstanden}
\alias{postcodes4_afstanden}
\title{Distance from Zipcode to Zipcode}
\format{
A \link{data.frame} with 562,330 observations and 3 variables:
\itemize{
\item \code{postcode.x}\cr zipcode (PC4)
\item \code{postcode.y}\cr zipcode (PC4)
\item \code{afstand_km}\cr distance in kilometres
}
}
\usage{
postcodes4_afstanden
}
\description{
This data set was obtained by calculating the difference from the middle point of a zipcode geometry to another zipcode geometry (using the \link{postcodes} data set and the \code{sf} package).
}
\keyword{datasets}
